[21] [PY] Функции. Программа сжатия изображений./requirements.txt
Pillow>=10.0.0
pillow-heif>=0.13.0
pillow-avif-plugin>=1.3.1
tqdm
